changed TicketID to Long

+++ TicketView/src/net/bechtelus/beans/JsfSearchByTicketID.java
@@ -24,7 +24,7 @@ public class JsfSearchByTicketID extends Ticket implements Serializable {
 
 	public String showTicket() {
 		if (!ticketID.isEmpty()) {
-			Faces.setSessionAttribute("ticketID", Integer.parseInt(ticketID));
+			Faces.setSessionAttribute("ticketID", Long.parseLong(ticketID));
 		}
 		return ("DspTicket"); // Means to go to page-b.xhtml (since condition
 								// is not mapped in faces-config.xml)
